requirements.txt
tensorflow==2.18.0
keras==3.8.0
streamlit==1.39.0
numpy>=1.26.0
pillow>=10.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
